feat: modify get expense route

+++ src/app/api/(protected)/group/[groupId]/expense/[expenseId]/route.ts
@@ -10,19 +10,12 @@ export async function GET(
     const expense = await prisma.expense.findUnique({
       where: { id: Number(params.expenseId), group: { id: params.groupId } },
       include: {
-        payer: {
-          select: {
-            id: true,
-            name: true,
-          },
-        },
         sharers: {
           select: {
             amount: true,
             user: {
               select: {
                 id: true,
-                name: true,
               },
             },
           },
@@ -30,19 +23,18 @@ export async function GET(
         historys: {
           select: {
             editedAt: true,
-            editor: {
-              select: {
-                id: true,
-                name: true,
-              },
-            },
+            editorId: true,
           },
         },
       },
     });
-
     if (!expense) return new NextResponse('Expense does not exist', { status: 404 });
-    return NextResponse.json(expense);
+
+    const responseBody: any = {
+      ...expense,
+      sharers: expense.sharers.map((sharer) => ({ id: sharer.user.id, amount: sharer.amount })),
+    };
+    return NextResponse.json(responseBody);
   } catch (error) {
     console.error(error);
     return NextResponse.error();
@@ -76,23 +68,29 @@ export async function PUT(
     return new NextResponse('Bad Request', { status: 400 });
   }
 
+  const group = await prisma.group.findUnique({ where: { id: params.groupId } });
+  if (!group) return new NextResponse('Group does not exist', { status: 404 });
+
   const expense = await prisma.expense.findUnique({
     where: { id: Number(params.expenseId), group: { id: params.groupId } },
   });
   if (!expense) return new NextResponse('Expense does not exist', { status: 404 });
-  // TODO: Should we first check whether payerId and sharerIds are valid?
+
+  const userIds = [
+    ...new Set([validatedBody.payerId, ...validatedBody.sharers.map((sharer) => sharer.id)]),
+  ];
+  const users = await prisma.groupUser.findMany({
+    where: { userId: { in: userIds } },
+  });
+  if (users.length !== userIds.length)
+    return new NextResponse('The user does not exist in the group', { status: 404 });
 
   try {
     const clientId = request.headers.get('client-id')!;
-
-    // TODO: Whether checking and updating can be put together?
     await prisma.sharer.deleteMany({
       where: { expense: { id: Number(params.expenseId) } },
     });
 
-    const e = await prisma.expense.findFirst({
-      where: { id: 1 },
-    });
     const expense = await prisma.expense.update({
       where: { id: Number(params.expenseId), group: { id: params.groupId } },
       data: {
@@ -114,19 +112,12 @@ export async function PUT(
         },
       },
       include: {
-        payer: {
-          select: {
-            id: true,
-            name: true,
-          },
-        },
         sharers: {
           select: {
             amount: true,
             user: {
               select: {
                 id: true,
-                name: true,
               },
             },
           },
@@ -134,18 +125,32 @@ export async function PUT(
         historys: {
           select: {
             editedAt: true,
-            editor: {
-              select: {
-                id: true,
-                name: true,
-              },
-            },
+            editorId: true,
           },
         },
       },
     });
+    const responseBody: any = {
+      ...expense,
+      sharers: expense.sharers.map((sharer) => ({ id: sharer.user.id, amount: sharer.amount })),
+    };
+    return NextResponse.json(responseBody);
+  } catch (error) {
+    console.error(error);
+    return NextResponse.error();
+  }
+}
 
-    return NextResponse.json(expense);
+export async function DELETE(
+  request: NextRequest,
+  { params }: { params: { groupId: string; expenseId: string } }
+) {
+  try {
+    const { count } = await prisma.expense.deleteMany({
+      where: { id: Number(params.expenseId), group: { id: params.groupId } },
+    });
+    if (count === 0) return new NextResponse('Expense does not exist', { status: 404 });
+    return new NextResponse(undefined, { status: 204 });
   } catch (error) {
     console.error(error);
     return NextResponse.error();
